move axios-request to main folder, fix tests
