+++ packages/insomnia/src/main/network/__tests__/axios-request.test.ts
@@ -1,8 +1,8 @@
 import { afterEach, beforeEach, describe, expect, it, jest } from '@jest/globals';
 import axios from 'axios';
 
-import { globalBeforeEach } from '../../__jest__/before-each';
-import * as models from '../../models';
+import { globalBeforeEach } from '../../../__jest__/before-each';
+import * as models from '../../../models';
 import { axiosRequest } from '../axios-request';
 
 interface AxiosRequestMockUserSettings {
