+++ packages/insomnia/src/main/network/axios-request.ts
@@ -3,9 +3,9 @@ import * as https from 'https';
 import { setDefaultProtocol } from 'insomnia-url';
 import { parse as urlParse } from 'url';
 
-import { isDevelopment } from '../common/constants';
-import * as models from '../models';
-import { isUrlMatchedInNoProxyRule } from './is-url-matched-in-no-proxy-rule';
+import { isDevelopment } from '../../common/constants';
+import * as models from '../../models';
+import { isUrlMatchedInNoProxyRule } from '../../network/is-url-matched-in-no-proxy-rule';
 
 export async function axiosRequest(config: AxiosRequestConfig) {
   const settings = await models.settings.getOrCreate();
@@ -54,6 +54,9 @@ export async function axiosRequest(config: AxiosRequestConfig) {
       res: {
         responseUrl: response?.request?.res?.responseUrl,
       },
+    },
+    config: {
+      proxy: response.config.proxy,
     },
   };
 }
